pyproject.toml
[project]
name = "BPP"
version = "0.1.0"
description = "Binding pocket prediction with EGNNs and Implicit Layers."
authors = [
    {name = "Darko Lukic", email = "darko.lukic@protonmail.com"},
]
dependencies = [
    "numpy>=1.23.5",
    "numba>=0.59.0",
    "torch<2.3.0",
    "torch-geometric>=2.4.0",
    "pytorch-lightning>=2.2.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.2",
    "biopython>=1.83",
    "graphein[extras]>=1.7.3",
    # "pymol @ git+https://github.com/schrodinger/pymol-open-source@v2.5.0",
    "rich>=13.7.1",
    "pdbx2df>=0.6.7",
    "biovec>=0.2.7",
    "wandb>=0.16.6",
    "typing-extensions>=4.6.0",
    "torchdeq>=0.1.0",
]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.dev-dependencies]
dev = [
    "ipython>=8.16.1",
    "pytest>=7.4.2",
    "hypothesis>=6.88.1",
    "ruff>=0.1.1",
    "pdoc>=14.4.0",
    "torchviz>=0.0.2",
]

[tool.pdm.resolution.overrides]
typing-extensions = ">=4.6.0"
